+++ spec/lib/aixm/horizontal/arc_spec.rb
@@ -25,7 +25,7 @@ describe AIXM::Horizontal::Arc do
         center_xy: AIXM::XY.new(lat: 22.2, long: 33.3),
         clockwise: true
       )
-      subject.to_digest.must_equal '35B2E1AF'
+      subject.to_digest.must_equal 712900173
     end
   end